move listen traits from io to server
