+++ io/src/net.rs
@@ -17,8 +17,7 @@ pub use tcp::{TcpListener, TcpStream};
 #[cfg(unix)]
 pub use unix::{UnixListener, UnixStream};
 
-use core::{future::Future, net::SocketAddr, pin::Pin};
-use std::io;
+use core::net::SocketAddr;
 
 macro_rules! default_aio_impl {
     ($ty: ty) => {
@@ -105,54 +104,6 @@ macro_rules! default_aio_impl {
 }
 
 use default_aio_impl;
-
-type BoxFuture<'f, T> = Pin<Box<dyn Future<Output = T> + Send + 'f>>;
-
-pub trait Listen: Send + Sync {
-    fn accept(&self) -> impl Future<Output = io::Result<Stream>> + Send;
-}
-
-#[doc(hidden)]
-pub trait ListenDyn: Send + Sync {
-    fn accept(&self) -> BoxFuture<io::Result<Stream>>;
-}
-
-impl<S> ListenDyn for S
-where
-    S: Listen,
-{
-    #[inline]
-    fn accept(&self) -> BoxFuture<io::Result<Stream>> {
-        Box::pin(Listen::accept(self))
-    }
-}
-
-impl Listen for TcpListener {
-    async fn accept(&self) -> io::Result<Stream> {
-        let (stream, addr) = self.accept().await?;
-        let stream = stream.into_std()?;
-        Ok(Stream::Tcp(stream, addr))
-    }
-}
-
-#[cfg(feature = "quic")]
-impl Listen for QuicListener {
-    async fn accept(&self) -> io::Result<Stream> {
-        let stream = self.accept().await?;
-        let addr = stream.peer_addr();
-        Ok(Stream::Udp(stream, addr))
-    }
-}
-
-#[cfg(unix)]
-impl Listen for UnixListener {
-    async fn accept(&self) -> io::Result<Stream> {
-        let (stream, _) = self.accept().await?;
-        let stream = stream.into_std()?;
-        let addr = stream.peer_addr()?;
-        Ok(Stream::Unix(stream, addr))
-    }
-}
 
 /// A collection of stream types of different protocol.
 #[allow(clippy::large_enum_variant)]
