+++ server/src/server/service.rs
@@ -1,14 +1,19 @@
-use std::{marker::PhantomData, rc::Rc, sync::Arc};
+use core::marker::PhantomData;
+
+use std::rc::Rc;
 
 use tokio::task::JoinHandle;
-use xitca_io::net::{ListenDyn, Stream};
+use xitca_io::net::Stream;
 use xitca_service::{Service, ready::ReadyService};
 
-use crate::worker::{self, ServiceAny};
+use crate::{
+    net::ListenerDyn,
+    worker::{self, ServiceAny},
+};
 
 pub type ServiceObj = Box<
     dyn for<'a> xitca_service::object::ServiceObject<
-            (&'a str, &'a [(String, Arc<dyn ListenDyn>)]),
+            (&'a str, &'a [(String, ListenerDyn)]),
             Response = (Vec<JoinHandle<()>>, ServiceAny),
             Error = (),
         > + Send
@@ -20,7 +25,7 @@ struct Container<F, Req> {
     _t: PhantomData<fn(Req)>,
 }
 
-impl<'a, F, Req> Service<(&'a str, &'a [(String, Arc<dyn ListenDyn>)])> for Container<F, Req>
+impl<'a, F, Req> Service<(&'a str, &'a [(String, ListenerDyn)])> for Container<F, Req>
 where
     F: IntoServiceObj<Req>,
     Req: TryFrom<Stream> + 'static,
@@ -30,7 +35,7 @@ where
 
     async fn call(
         &self,
-        (name, listeners): (&'a str, &'a [(String, Arc<dyn ListenDyn>)]),
+        (name, listeners): (&'a str, &'a [(String, ListenerDyn)]),
     ) -> Result<Self::Response, Self::Error> {
         let service = self.inner.call(()).await.map_err(|_| ())?;
         let service = Rc::new(service);
